Fix argparse bug

+++ train.py
@@ -10,7 +10,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--d_model", type=int)
     parser.add_argument("--data_len", type=int)
-    parser.add_argument("--dropout", type=int)
+    parser.add_argument("--dropout", type=float)
     parser.add_argument("--ff", type=int)
     parser.add_argument("--nhead", type=int)
     parser.add_argument("--num_layers", type=int)
